api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
redis==5.0.2
